move parallel map serialization tests to kernel tests
